pyproject.toml
[project]
name = "dexa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.13",
    "langgraph>=1.1.3",
    "pandas>=3.0.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "typer>=0.24.1",
    "langchain-groq>=0.1.6",
    "langchain-core>=0.2.21",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "kaggle>=1.6.0",
    "pyarrow>=15.0.0"
]
